+++ actionpack/lib/action_controller/metal/live.rb
@@ -39,22 +39,6 @@ module ActionController
   # the main thread. Make sure your actions are thread safe, and this shouldn't be
   # a problem (don't share state across threads, etc).
   #
-  # Note that Rails includes `Rack::ETag` by default, which will buffer your
-  # response. As a result, streaming responses may not work properly with Rack
-  # 2.2.x, and you may need to implement workarounds in your application. You can
-  # either set the `ETag` or `Last-Modified` response headers or remove
-  # `Rack::ETag` from the middleware stack to address this issue.
-  #
-  # Here's an example of how you can set the `Last-Modified` header if your Rack
-  # version is 2.2.x:
-  #
-  #     def stream
-  #       response.headers["Content-Type"] = "text/event-stream"
-  #       # Add this line if your Rack version is 2.2.x
-  #       response.headers["Last-Modified"] = Time.now.httpdate
-  #       ...
-  #     end
-  #
   # ## Streaming and Execution State
   #
   # When streaming, the action is executed in a separate thread. By default, this thread