+++ railties/lib/rails/rackup/server.rb
@@ -4,12 +4,7 @@
 
 module Rails
   module Rackup
-    begin
-      require "rackup/server"
-      Server = ::Rackup::Server
-    rescue LoadError
-      require "rack/server"
-      Server = ::Rack::Server
-    end
+    require "rackup/server"
+    Server = ::Rackup::Server
   end
 end